# docs/openapi.json
{
  "components": {
    "schemas": {
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PlatformRole": {
        "enum": [
          "attendee",
          "platform_admin"
        ],
        "title": "PlatformRole",
        "type": "string"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "locale": {
            "default": "ru",
            "enum": [
              "kk",
              "ru",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "ResendVerificationRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ResendVerificationRequest",
        "type": "object"
      },
      "ResetPasswordRequest": {
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "description": "The `token` query parameter from the link.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "description": "JWT for `Authorization: Bearer ...`.",
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Access token lifetime in seconds.",
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "description": "Single-use: exchange it at `POST /auth/refresh` for a new pair.",
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "refresh_token",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "description": "True once the email address has been verified.",
            "title": "Email Verified",
            "type": "boolean"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locale": {
            "enum": [
              "kk",
              "ru",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "platform_role": {
            "$ref": "#/components/schemas/PlatformRole"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "phone",
          "locale",
          "platform_role",
          "email_verified",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "token": {
            "description": "The `token` query parameter from the link.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "VerifyEmailRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from `POST /auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BiletFlow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Returns a 15-minute access token (send as `Authorization: Bearer ...`) and a single-use refresh token. Unverified accounts can sign in; actions that need a verified email return 403 `EMAIL_NOT_VERIFIED`. Errors: 401 `INVALID_CREDENTIALS`, 403 `ACCOUNT_SUSPENDED`.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revokes the given refresh token (this device only). Always returns 204. The access token stays valid until it expires, so clients should discard it.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign out",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "description": "Emails a single-use reset link (`/reset-password?token=...` on the web app, valid 60 minutes) and invalidates older links. Always returns 202, whether or not the email is registered; repeated requests within a short cooldown send nothing new.",
        "operationId": "forgot_password_api_v1_auth_password_forgot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a password reset email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "description": "Sets the new password, marks the email verified, and signs out every session (all refresh tokens revoked). Sign in again afterwards. Errors: 400 `RESET_TOKEN_EXPIRED` (request a new link), 400 `RESET_TOKEN_INVALID` (unknown, already used, or superseded).",
        "operationId": "reset_password_api_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set a new password with a reset link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Exchanges a refresh token for a new access token and a new refresh token; the old refresh token stops working (rotation). Call it once per expiry: a second call with the same token fails. Errors: 401 `REFRESH_TOKEN_INVALID` (sign in again), 403 `ACCOUNT_SUSPENDED`.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh the access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates an unverified attendee account and queues a verification email. Returns 409 `EMAIL_TAKEN` if the email (case-insensitive) is already registered. Does not sign the user in; use login afterwards.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a new account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Redeems the token from the verification link (`/verify-email?token=...` on the web app) and returns the verified user. Posting an already-redeemed token for a verified user returns 200 again. Errors: 400 `VERIFICATION_TOKEN_EXPIRED` (offer a resend), 400 `VERIFICATION_TOKEN_INVALID` (unknown, or superseded by a newer link).",
        "operationId": "verify_email_api_v1_auth_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify an email address",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verify-email/resend": {
      "post": {
        "description": "Queues a new verification link for an unverified account and invalidates older links. Always returns 202, whether or not the email is registered or already verified; repeated requests within a short cooldown send nothing new.",
        "operationId": "resend_verification_api_v1_auth_verify_email_resend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend the verification email",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}